+++ src/fobi/contrib/plugins/form_handlers/mail/forms.py
@@ -10,6 +10,8 @@ from django import forms
 from django.utils.translation import ugettext_lazy as _
 
 from fobi.base import BasePluginForm, get_theme
+from fobi.contrib.plugins.form_handlers.mail.fields import MultiEmailField
+from fobi.contrib.plugins.form_handlers.mail.widgets import MultiEmailWidget
 
 theme = get_theme(request=None, as_instance=True)
 
@@ -29,30 +31,42 @@ class MailForm(forms.Form, BasePluginForm):
     from_name = forms.CharField(
         label = _("From name"),
         required = True,
-        widget = forms.widgets.TextInput(attrs={'class': theme.form_element_html_class})
+        widget = forms.widgets.TextInput(
+            attrs={'class': theme.form_element_html_class}
+            )
         )
     from_email = forms.EmailField(
         label = _("From email"),
         required = True,
-        widget = forms.widgets.TextInput(attrs={'class': theme.form_element_html_class})
+        widget = forms.widgets.TextInput(
+            attrs={'class': theme.form_element_html_class}
+            )
         )
     to_name = forms.CharField(
         label = _("To name"),
         required = True,
-        widget = forms.widgets.TextInput(attrs={'class': theme.form_element_html_class})
+        widget = forms.widgets.TextInput(
+            attrs={'class': theme.form_element_html_class}
+            )
         )
-    to_email = forms.EmailField(
+    to_email = MultiEmailField(#forms.EmailField(
         label = _("To email"),
         required = True,
-        widget = forms.widgets.TextInput(attrs={'class': theme.form_element_html_class})
+        widget = MultiEmailWidget(#forms.widgets.TextInput(
+            attrs={'class': theme.form_element_html_class}
+            )
         )
     subject = forms.CharField(
         label = _("Subject"),
         required = True,
-        widget = forms.widgets.TextInput(attrs={'class': theme.form_element_html_class})
+        widget = forms.widgets.TextInput(
+            attrs={'class': theme.form_element_html_class}
+            )
         )
     body = forms.CharField(
         label = _("Body"),
         required = False,
-        widget = forms.widgets.Textarea(attrs={'class': theme.form_element_html_class})
+        widget = forms.widgets.Textarea(
+            attrs={'class': theme.form_element_html_class}
+            )
         )