+++ src/fobi/contrib/themes/bootstrap3/widgets/form_elements/datetime_bootstrap3_widget/static/bootstrap3/js/fobi.plugin.datetime-bootstrap3-widget.js
@@ -1,3 +1,10 @@
+/*
+    Document   : fobi.plugin.datetime-bootstrap3-widget.js
+    Created on : Apr 26, 2015, 03:30:35 AM
+    Author     : Artur Barseghyan (artur.barseghyan@gmail.com)
+    Description:
+        `bootstrap3` theme datetime picker plugin scripts.
+*/
 ;
 
 $(document).ready(function() {