+++ src/fobi/contrib/themes/foundation5/widgets/form_elements/date_foundation5_widget/__init__.py
@@ -1,0 +1,7 @@
+__title__ = 'fobi.contrib.themes.foundation5.widgets.form_elements.date_foundation5_widget'
+__author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
+__copyright__ = 'Copyright (c) 2014-2015 Artur Barseghyan'
+__license__ = 'GPL 2.0/LGPL 2.1'
+__all__ = ('default_app_config',)
+
+default_app_config = 'fobi.contrib.themes.foundation5.widgets.form_elements.date_foundation5_widget.apps.Config'