+++ src/fobi/contrib/themes/foundation5/widgets/form_elements/dummy_foundation5_widget/fobi_form_elements.py
@@ -1,18 +1,18 @@
-__title__ = 'fobi.contrib.themes.bootstrap3.widgets.form_elements.dummy_bootstrap3_widget.fobi_form_elements'
+__title__ = 'fobi.contrib.themes.foundation5.widgets.form_elements.dummy_foundation5_widget.fobi_form_elements'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 __copyright__ = 'Copyright (c) 2014-2015 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = ('DummyPluginWidget',)
 
 from fobi.base import form_element_plugin_widget_registry
-from fobi.contrib.themes.bootstrap3 import UID
+from fobi.contrib.themes.foundation5 import UID
 from fobi.contrib.plugins.form_elements.test.dummy.widgets import (
     BaseDummyPluginWidget
 )
 
 class DummyPluginWidget(BaseDummyPluginWidget):
     """
-    Dummy plugin widget for Boootstrap 3.
+    Dummy plugin widget for Foundation 5.
     """
     theme_uid = UID
     media_js = [